Move ConnectionAdmin to connections folder
